+++ clickhouse_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit clickhouse.stub.php instead.
- * Stub hash: 2fa3a1f4da0cefd8b230a7d6cd155ea30dd2660a */
+ * Stub hash: 34382e092fbf02658c9f9d33116ac2218edc37bd */
 
 ZEND_BEGIN_ARG_INFO_EX(arginfo_class_ClickHouse___construct, 0, 0, 1)
 	ZEND_ARG_TYPE_INFO(0, connectParams, IS_ARRAY, 0)
@@ -21,6 +21,15 @@ ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_ClickHouse_selectWithExter
 	ZEND_ARG_TYPE_INFO(0, externals, IS_ARRAY, 0)
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, params, IS_ARRAY, 0, "[]")
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, fetch_mode, IS_LONG, 0, "0")
+	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, query_id, IS_STRING, 0, "\"\"")
+	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, settings, IS_ARRAY, 0, "[]")
+ZEND_END_ARG_INFO()
+
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_ClickHouse_selectToStream, 0, 3, IS_LONG, 0)
+	ZEND_ARG_TYPE_INFO(0, sql, IS_STRING, 0)
+	ZEND_ARG_TYPE_INFO(0, params, IS_ARRAY, 0)
+	ZEND_ARG_TYPE_INFO(0, stream, IS_MIXED, 0)
+	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, format, IS_STRING, 0, "\"TabSeparated\"")
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, query_id, IS_STRING, 0, "\"\"")
 	ZEND_ARG_TYPE_INFO_WITH_DEFAULT_VALUE(0, settings, IS_ARRAY, 0, "[]")
 ZEND_END_ARG_INFO()
@@ -237,6 +246,7 @@ ZEND_METHOD(ClickHouse, __construct);
 ZEND_METHOD(ClickHouse, __destruct);
 ZEND_METHOD(ClickHouse, select);
 ZEND_METHOD(ClickHouse, selectWithExternalData);
+ZEND_METHOD(ClickHouse, selectToStream);
 ZEND_METHOD(ClickHouse, insert);
 ZEND_METHOD(ClickHouse, insertAssoc);
 ZEND_METHOD(ClickHouse, writeStart);
@@ -304,6 +314,7 @@ static const zend_function_entry class_ClickHouse_methods[] = {
 	ZEND_ME(ClickHouse, __destruct, arginfo_class_ClickHouse___destruct, ZEND_ACC_PUBLIC)
 	ZEND_ME(ClickHouse, select, arginfo_class_ClickHouse_select, ZEND_ACC_PUBLIC)
 	ZEND_ME(ClickHouse, selectWithExternalData, arginfo_class_ClickHouse_selectWithExternalData, ZEND_ACC_PUBLIC)
+	ZEND_ME(ClickHouse, selectToStream, arginfo_class_ClickHouse_selectToStream, ZEND_ACC_PUBLIC)
 	ZEND_ME(ClickHouse, insert, arginfo_class_ClickHouse_insert, ZEND_ACC_PUBLIC)
 	ZEND_ME(ClickHouse, insertAssoc, arginfo_class_ClickHouse_insertAssoc, ZEND_ACC_PUBLIC)
 	ZEND_ME(ClickHouse, writeStart, arginfo_class_ClickHouse_writeStart, ZEND_ACC_PUBLIC)